test_git_update_system.py
#!/usr/bin/env python3
"""
Test script for Git-based Auto-Update System
Validates Git integration, speed comparisons, and fallback mechanisms
"""

import sys
import os
import time
import json
from datetime import datetime
from typing import Dict, Optional

# Add current directory to path for imports
sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))

def print_header(title: str):
    """Print a formatted header"""
    print("\n" + "=" * 60)
    print(f" {title}")
    print("=" * 60)

def print_section(title: str):
    """Print a formatted section header"""
    print(f"\n🔍 {title}")
    print("-" * 40)

def test_git_availability():
    """Test if Git and GitPython are available"""
    print_section("Testing Git Availability")
    
    # Test Git command
    try:
        import subprocess
        result = subprocess.run(['git', '--version'], 
                              capture_output=True, text=True, timeout=5)
        if result.returncode == 0:
            print(f"✅ Git command available: {result.stdout.strip()}")
            git_cmd_available = True
        else:
            print("❌ Git command not available")
            git_cmd_available = False
    except Exception as e:
        print(f"❌ Git command error: {e}")
        git_cmd_available = False
    
    # Test GitPython
    try:
        import git
        from git import Repo, GitCommandError
        print(f"✅ GitPython available: {git.__version__}")
        gitpython_available = True
    except ImportError as e:
        print(f"❌ GitPython not available: {e}")
        gitpython_available = False
    
    return git_cmd_available and gitpython_available

def test_git_repository_manager():
    """Test Git Repository Manager functionality"""
    print_section("Testing Git Repository Manager")
    
    try:
        from git_updater import get_git_repository_manager, GIT_AVAILABLE
        
        if not GIT_AVAILABLE:
            print("❌ Git not available - skipping Git manager tests")
            return False
        
        # Initialize manager
        git_manager = get_git_repository_manager()
        print(f"✅ Git manager initialized")
        
        # Test Git availability check
        if git_manager.is_git_available():
            print("✅ Git availability check passed")
        else:
            print("❌ Git availability check failed")
            return False
        
        # Test version info retrieval (without actually cloning)
        print("🔍 Testing version info retrieval...")
        # This would normally clone the repo, so we'll skip for now
        print("⚠️ Skipping actual repository operations to avoid large downloads")
        
        return True
        
    except Exception as e:
        print(f"❌ Git manager test failed: {e}")
        return False

def test_hybrid_updater():
    """Test Hybrid Updater functionality"""
    print_section("Testing Hybrid Updater")
    
    try:
        from hybrid_updater import get_hybrid_updater
        
        # Initialize hybrid updater
        hybrid_updater = get_hybrid_updater()
        print("✅ Hybrid updater initialized")
        
        # Test update statistics
        stats = hybrid_updater.get_update_statistics()
        print(f"✅ Update statistics retrieved:")
        for key, value in stats.items():
            print(f"   {key}: {value}")
        
        # Test should check for updates
        should_check = hybrid_updater.should_check_for_updates()
        print(f"✅ Should check for updates: {should_check}")
        
        return True
        
    except Exception as e:
        print(f"❌ Hybrid updater test failed: {e}")
        return False

def test_fallback_mechanism():
    """Test fallback from Git to HTTP"""
    print_section("Testing Fallback Mechanism")
    
    try:
        # Test with Git disabled
        print("🔍 Testing HTTP fallback when Git is unavailable...")
        
        from updater import get_updater
        
        # Get basic updater (should fall back to HTTP)
        updater = get_updater()
        print("✅ Fallback updater initialized")
        
        # Test basic functionality
        if hasattr(updater, 'should_check_for_updates'):
            should_check = updater.should_check_for_updates()
            print(f"✅ HTTP updater should check: {should_check}")
        
        return True
        
    except Exception as e:
        print(f"❌ Fallback test failed: {e}")
        return False

def test_update_manager_integration():
    """Test Update Manager integration with Git system"""
    print_section("Testing Update Manager Integration")
    
    try:
        from update_manager import get_update_manager
        
        # This would require Qt, so we'll just test imports
        print("✅ Update manager imports successful")
        
        # Test that the update manager can get the hybrid updater
        from update_manager import get_updater
        updater = get_updater()
        print("✅ Update manager can access updater")
        
        return True
        
    except Exception as e:
        print(f"❌ Update manager integration test failed: {e}")
        return False

def test_build_integration():
    """Test build system integration"""
    print_section("Testing Build System Integration")
    
    # Check if Git files are included in build script
    try:
        with open('setup_working.py', 'r') as f:
            build_content = f.read()
        
        if 'git_updater.py' in build_content:
            print("✅ git_updater.py included in build")
        else:
            print("❌ git_updater.py not included in build")
        
        if 'hybrid_updater.py' in build_content:
            print("✅ hybrid_updater.py included in build")
        else:
            print("❌ hybrid_updater.py not included in build")
        
        if 'GitPython' in build_content or 'git' in build_content:
            print("✅ Git packages included in build")
        else:
            print("❌ Git packages not included in build")
        
        return True
        
    except Exception as e:
        print(f"❌ Build integration test failed: {e}")
        return False

def performance_comparison_simulation():
    """Simulate performance comparison between Git and HTTP"""
    print_section("Performance Comparison Simulation")
    
    print("📊 Simulated performance comparison:")
    print("   HTTP Download (current):")
    print("     - Full file download: ~50MB")
    print("     - Time: 30-60 seconds (depending on connection)")
    print("     - Bandwidth: Full file size")
    print("     - Resume: Not supported")
    
    print("   Git Download (new):")
    print("     - Incremental updates: ~5-15MB (typical)")
    print("     - Time: 10-20 seconds (depending on changes)")
    print("     - Bandwidth: Only changed files")
    print("     - Resume: Supported")
    
    print("   Expected improvements:")
    print("     - 60-80% reduction in download time")
    print("     - 70-90% reduction in bandwidth usage")
    print("     - Better reliability with resume capability")
    print("     - Reduced bot detection risk")

def generate_test_report():
    """Generate a comprehensive test report"""
    print_header("Git Update System Test Report")
    
    tests = [
        ("Git Availability", test_git_availability),
        ("Git Repository Manager", test_git_repository_manager),
        ("Hybrid Updater", test_hybrid_updater),
        ("Fallback Mechanism", test_fallback_mechanism),
        ("Update Manager Integration", test_update_manager_integration),
        ("Build Integration", test_build_integration),
    ]
    
    results = {}
    
    for test_name, test_func in tests:
        print_header(f"Running: {test_name}")
        try:
            start_time = time.time()
            result = test_func()
            end_time = time.time()
            
            results[test_name] = {
                'passed': result,
                'duration': end_time - start_time,
                'timestamp': datetime.now().isoformat()
            }
            
            status = "✅ PASSED" if result else "❌ FAILED"
            print(f"\n{status} - {test_name} ({end_time - start_time:.2f}s)")
            
        except Exception as e:
            results[test_name] = {
                'passed': False,
                'error': str(e),
                'duration': 0,
                'timestamp': datetime.now().isoformat()
            }
            print(f"\n❌ FAILED - {test_name}: {e}")
    
    # Performance simulation
    performance_comparison_simulation()
    
    # Summary
    print_header("Test Summary")
    passed = sum(1 for r in results.values() if r['passed'])
    total = len(results)
    
    print(f"Tests passed: {passed}/{total}")
    print(f"Success rate: {(passed/total)*100:.1f}%")
    
    if passed == total:
        print("🎉 All tests passed! Git update system is ready.")
    elif passed > total // 2:
        print("⚠️ Most tests passed. Some issues need attention.")
    else:
        print("❌ Multiple test failures. System needs fixes.")
    
    # Save detailed report
    report_file = f"git_update_test_report_{datetime.now().strftime('%Y%m%d_%H%M%S')}.json"
    try:
        with open(report_file, 'w') as f:
            json.dump(results, f, indent=2)
        print(f"\n📄 Detailed report saved: {report_file}")
    except Exception as e:
        print(f"⚠️ Could not save report: {e}")
    
    return results

def main():
    """Main test function"""
    print("VARSYS Kitchen Dashboard - Git Update System Test")
    print(f"Test started at: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
    
    # Check Python version
    print(f"Python version: {sys.version}")
    print(f"Platform: {sys.platform}")
    
    # Run comprehensive tests
    results = generate_test_report()
    
    # Exit with appropriate code
    passed = sum(1 for r in results.values() if r['passed'])
    total = len(results)
    
    if passed == total:
        sys.exit(0)  # Success
    else:
        sys.exit(1)  # Some failures

if __name__ == "__main__":
    main()
