requirements.txt
# Core dependencies
pandas>=1.5.0
matplotlib>=3.5.0
openpyxl>=3.0.0
python-dotenv>=1.0.0
pillow>=9.0.0
PySide6>=6.0.0
# pyinstaller>=5.8.0  # Removed due to circular dependency issues

# Firebase integration (enabled for v1.1.1)
firebase-admin>=6.0.0
pyrebase4>=4.5.0

# Data handling and visualization
numpy>=1.22.0
seaborn>=0.12.0

# Utility libraries
tqdm>=4.64.0
requests>=2.28.0
json5>=0.9.10
loguru>=0.6.0

# Date handling
python-dateutil>=2.8.2

# AI & Machine Learning
scikit-learn>=1.3.0
PyJWT>=2.8.0

# Auto-update system dependencies
urllib3>=1.26.0
certifi>=2022.12.7

# Git-based update system (optional for faster downloads)
GitPython>=3.1.40

# License management and security
cryptography>=41.0.0

# Build and distribution
setuptools>=65.0.0
wheel>=0.38.0
cx_Freeze>=6.15.0

# Windows-specific dependencies for EXE compatibility
pywin32>=305; sys_platform == "win32"
pywin32-ctypes>=0.2.0; sys_platform == "win32"
